Move artists/songs to their own controller
